requirements.txt
# Core Testing Framework
behave>=1.2.6

# Web Application
flask>=3.0.0

# Web Automation
selenium>=4.0.0

# HTML Parsing (used by test runner for element discovery)
beautifulsoup4>=4.12.0

# LLM Integration (Gemini API for NL step parsing and scenario generation)
google-genai>=1.0.0

# API Testing
requests>=2.28.0

# Additional Testing Tools
pytest>=7.0.0
pytest-cov>=4.0.0

# Optional: Enhanced Reporting
allure-behave>=2.13.0

# Optional: Environment Configuration
python-dotenv>=1.0.0

# Screenshot generation (used by simulated execution to create step images)
Pillow>=10.0.0

# Optional: Assertions
assertpy>=1.1
